게시물 테이블 구조 변경, memberId 추가, doWrite시 로그인 체크, 작성자 정보 저장

+++ src/main/java/com/KoreaIT/sdy/demo/controller/UsrArticleController.java
@@ -2,6 +2,8 @@ package com.KoreaIT.sdy.demo.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -39,7 +41,19 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = -1;
+		
+		if(httpSession.getAttribute("loginedMemberId")!=null) {
+			isLogined = true;
+			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (isLogined==false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
 		if(Ut.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요.");
 		}
@@ -47,7 +61,7 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요.");
 		}
 		
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body, loginedMemberId);
 		
 		int id = (int) writeArticleRd.getData1();
 		
@@ -58,31 +72,61 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int loginedMemberId = -1;
+		
+		if(httpSession.getAttribute("loginedMemberId")!=null) {
+			isLogined = true;
+			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (isLogined==false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		if(article==null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 존재하지 않습니다.", id)); 
 		}
-		else {
-			articleService.deleteArticle(id);
-			return ResultData.from("S-1", Ut.f("%d번 게시글이 삭제되었습니다.", id), id); 
+		
+		if(loginedMemberId==article.getMemberId()==false) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 권한이 없습니다.", id)); 
 		}
+		
+		articleService.deleteArticle(id);
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 삭제되었습니다.", id), id); 
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Article> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = -1;
+		
+		if(httpSession.getAttribute("loginedMemberId")!=null) {
+			isLogined = true;
+			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if (isLogined==false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		if(article==null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시글은 존재하지 않습니다.", id)); 
 		}
-		else {
-			articleService.modifyArticle(id, title, body);
-			article = articleService.getArticleById(id);
-			return ResultData.from("S-1", Ut.f("%d번 게시글이 수정되었습니다.", id), article); 
+		
+		if(loginedMemberId==article.getMemberId()==false) {
+			return ResultData.from("F-2", Ut.f("%d번 게시글에 권한이 없습니다.", id)); 
 		}
+		
+		articleService.modifyArticle(id, title, body);
+		article = articleService.getArticleById(id);
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 수정되었습니다.", id), article); 
 	}
 
 }
